Fix failing on KDoc semicolon after other nodes.
/**
 * x;
 */
was working fine because the text was processed as "x;" but
/**
 * [x];
 */
failed because the left bracket, the identifier, and the right racket were processed as separate nodes, and then the node with the single ";" text was checked and failed.
Simply opting out of the check for all KDOC_TEXT seems fine.