Add Angle Brackets Rule
Adds a rule to lint against white space before and after angle brackets in type arguments. This includes all of the following:
Map< String, Int>
Map<String, Int >
Map <String, Int>

In all of the above cases, the desired format is
Map<String, Int>

Specifically this rule does not affect generics formatting of the type
fun myFun<T>()
as this is already implemented in a separate rule.