Add experimental SpacingAroundAngleBracketsRule (#769)
* Move angle bracket test to experimental heirarchy

The new angle bracket rule will live in the experimental rule set along with all the other new rules, so this commit refactors the associated tests to also live under the experimental package

* Add Angle Brackets Rule

Adds a rule to lint against white space before and after angle brackets in type arguments. This includes all of the following:
Map< String, Int>
Map<String, Int >
Map <String, Int>

In all of the above cases, the desired format is
Map<String, Int>

Specifically this rule does not affect generics formatting of the type
fun myFun<T>()
as this is already implemented in a separate rule.

* Handle newlines and TYPE_PARAMETER_LIST

* Move angle bracket test to experimental heirarchy

The new angle bracket rule will live in the experimental rule set along with all the other new rules, so this commit refactors the associated tests to also live under the experimental package

* Add Angle Brackets Rule

Adds a rule to lint against white space before and after angle brackets in type arguments. This includes all of the following:
Map< String, Int>
Map<String, Int >
Map <String, Int>

In all of the above cases, the desired format is
Map<String, Int>

Specifically this rule does not affect generics formatting of the type
fun myFun<T>()
as this is already implemented in a separate rule.

* Handle newlines and TYPE_PARAMETER_LIST

* Rename to spacing-around-angle-brackets and fix "fun"

* Add more tests

Co-authored-by: Christina Lee <lee@lee-5CPJGH6.dyn.pinadmin.com>
Co-authored-by: Yash Mittal <ymittal@pinterest.com>