Opening class RuleSet so it can be inherited.
### What's done:
In some cases some custom rule sets need to have different implementation of RuleSet. For example in the diktat (https://github.com/cqfn/diKTat) project it is used to pass ruleset specific configuration in the rules (config files). This should not affect anything and simply gives developers a chance to customise rule sets.