Update README.md
When pasting this to `pom.xml` it corrupts it. Because, it's not allowed to have `--` inside of comments.
You can paste it, for example, here: https://www.xmlvalidation.com/ to check.

After pasting the suggested snippet Maven stops working, I don't see another way to cure it, let's just remove it then.