Adding new option disable-standard-rules (#977)
* Adding new option disable-standard-rules

### What's done:
Adding special option that is extremely useful and needed when ktlint is used as a framework for static analysis. In some cases when custom ruleset is used - there is no need to use standard ruleset as it can lead to conflicts.

* Changing option --disabled_rules to have a 'standard' mode that disables standard ruleset

### What's done:
Added functionality that allows to have --disabled_rules=standard. This will let you disable standard set of rules in ktlint. Should be useful if you are using some custom set of rules.

* Changing option --disabled_rules to have a 'standard' mode that disables standard ruleset

### What's done:
Fixing CI/CD

* Changing option --disabled_rules to have a 'standard' mode that disables standard ruleset

### What's done:
Fixing CI/CD