Move rule initialization out of normal visit of node
Make a clearer distinction between the initialization of a rule
versus processing (e.g. visiting) the rule with a node. A rule
that implements the Rule.Modifier.InitializeRoot interface will
be called with the rootNode to initialize the rule. After that
the rule will be visited for each node (including the root
node) unless the node is suppressed by a ktlint-disable
directive.