Simplify and improve check on suppressed regions
When starting a new block in which a rule is disabled, then define the
range of this block from the startOffset of that block until eternity
(Int.MAX_VALUE). When the block is closed explicitly by a ktlint-enable
directive, or when closing all open blocks at the end of the collect
phase, the end position of the block is overwritten with the offset of
the end of the file.

This fixes a very specific bug in the IndentationRule which on occasion
modifies the structure of the ASTNodes by inserting new nodes. As of
such, the rootNode can contain offsets which are not equal to zero but
which still need to be suppressed. This bug can not be tested in the
unit test of the SupressedRegionLocator but is verified in the unit
test IndentationRuleTest.