Extract VisitorProvider from Ktlint class
The visitor is created once for each file that has to be linted/formatted. It is
now divided into two distinct phases.

The setup phase is executed whenever the VisitorProvider is created. This provider
is created only once per invocation of the KtLint command. In the setup phase, the
rules are analysed and the order in which they have to be executed is determined.
This order is identical for all nodes. As the phase is executed only once, it can
now containing logging about the order in which the rules are executed. Also it
opens up possibilities for more complex ordering rules using annotations and
reflection (for example executing a rule after a specific rule).

The visitor creation phase actually creates the visitor. In this phase all enabled
rules will be executed. Note that disabling of rules depends on the
'.editorconfig' which is active for a given rootNode.