Move sorting on ruleSetId from RuleSetsLoader to VisitorProvider
The order in which the rule sets are discovered is not relevant for the
loader. Ordering rules by rule set however is important within the scope
of the VisitorProvider though.

Also add order that experimental rules takes precedence over custom rules.
Finally add secondary sorting on ruleId so that the order of the rules
as defined by the RuleSetProviders is not relevant anymore.