Restrict usage of Rule.Modifier to initialisation phase in which the order of the rules is determined
Rule.Modifier is no longer used during actual creation of the visitor. When creating the
visitor the rules are already ordered. This opens up the new possibilities to add other
rule ordering mechanisms.

Note that this introduces *functional* changes below with regard to the order of
execution of the rules:
 - Rules without Rule.Modifier and rules with Rule.Modifier.RestrictToRoot are now
   handled as one group instead of two distinct groups.
 - Rules with Rule.Modifier.Last and rules with Rule.Modifier.RestrictToRootLast are
   now handled as one group instead of two distinct groups.