Fix indentation of the closing quotes of a multiline string literal
Closes #1127

Arranging the opening and closing quotes of multiline string literals should not be
limited to strings followed by the ".trimIndent()" clause.

In case the opening quotes have been placed at the start of the line, then the
closing quotes should also be placed at the start of the line.

In the reporter tests, the tab characters have replaced with an explicit placeholder
"{TAB}" to make them visible.