Do not remove trailing comma in value argument list in annotation
The code previously contained special handling for case below:
    @Annotation([
        "something",
    ])

This special case however was also triggered in case of:
    @Annotation(
       foo = [
        "something"
       ],
       bar = "some-bar"
   )
In this case the normal handling of value argument lists should be
used once for the outer value argument list (containing elements foo
and bar) and once for the inner value argument list of parameter foo
(containing element "something"). So in example above, a trailing
comma has to be added to both the outer and the inner value argument
list.

Closes #1297