Fix tests regarding warning when a provided JAR does not contain a ruleset
Note: The jar "ktlint-ruleset-experimental.jar" which previously was used
as invalid jar in the unit test, was replaced in the master branch with a
jar that actually does contain a ruleset provider. Now a new JAR (the
"ktlint-reporter-html.jar") is added which does not contain a ruleset
provider. Also the tests have been changed to check in the normal output
for absence or existence of the warning.