Change log level of messages regarding which file is being processed and what .editorconfig value are applicable from debug to trace
Loglevel debug is enabled by the verbose mode which is commonly used on the CLI to detect
which rule actually is throwing the lint error. This same flag is also used by the build
step which checks the ktlint source code for violations. Due to way too many errors, it
became too hard to find the files/lines which need to be changed.