A single line function with a parameter having a lambda as default argument should not throw error (#1358)
A single line function with a parameter having a lambda as default argument should not throw error

Making this possible required a considerable refactoring of the code.
The exceptions regarding formatting of blocks in conjunction with
the exceptions for the BY-keyword continually conflicted with each
other. Especially the exceptions for the BY-keyword were sometimes
incorrectly implemented.

The blockOpeningLineStack in the IndentContext has now been replaced
with a blockStack which contains more context information about the
block. Also trace logging now provides more details about opening
and closing blocks. The main idea behind this stack is however not
changed. But some functionality has been moved from the closing
element to the opening element.

Closes #1330

Co-authored-by: Paul Dingemans <pdingemans@bol.com>
Co-authored-by: Sha Sha Chu <shasha@pinterest.com>