When running format mode, the rules are first executed in parallel to… (#1368)
* When running format mode, the rules are first executed in parallel to find linting errors. In this process,
no unused import are found because the trailing comma is not yet added to variable "bar3". Then in the next
stage the rules are run consecutively. Nof the trailing comma rule is adding a trailing comma after the type
of variable "bar3". When the no-unused-import rule runs after the trailing-comma rule, it was incorrectly
seen as part of the type of variable "bar3" and a reference "EnumThree," (with the trailing comma was added)
which in turn resulted in not recognizing that the import of EnumThree actually was used.

Closes #1367

* Update changelog

Co-authored-by: Paul Dingemans <pdingemans@bol.com>
Co-authored-by: Sha Sha Chu <shasha@pinterest.com>