Add EditorConfigOverride for unit tests
Currently, an unit test has to generate an ".editorconfig" file in
a temporary folder when the test needs to set a specific value for
and editor config property. For this the JUnit4 Rule construct is
used which has no equivalent in JUnit5. Also, it is considered to
be a code smell when a unit tests needs access to the file system.

In order to make this work, method loadPropertiesForFile in the
EditorConfigLoader has been fixed. Now, it does return the
EditorConfigOverridesMap instead of an emptyMap() when an unit
test runs.

New RuleExtension functions have been added to RuleExtension
which support the EditorConfigOverride parameter. Other
functions are added so that no useless deprecation
warning are shown in case that the function invocation
does not specify the userData parameter.
Old functions have been deprecated.