Add rule to report discouraged comment locations (#1365)
Autocorrect is not implemented. This is an error which only
happens rarely. Its value is that it makes implementation of
some rules way easier when comment do not have to be taken
into account.

Initial implementation just contains one single discouraged
location. More will follow when implementing the rule to
rewrite the function signature automatically as is described
in #1341.