Add new experimental rule "function-signature"
This rule rewrites the function signature to a single line when possible (e.g. when not exceeding the "max_line_length" property) or a multiline signature otherwise. In case of function with a body expression, the body expression is placed on the same line as the function signature when not exceeding the "max_line_length" property. Optionally the function signature can be forced to be written as a multiline signature in case the function has more than a specified number of parameters (".editorconfig" property "ktlint_function_signature_wrapping_rule_always_with_minimum_parameters")

Fixed the FunctionStartOfBodySpacingRule to add the missing space before start of the block body before the BLOCK element type instead of as first element inside the BLOCK element.