Allow rules to stop the traversal of the AST whenever the goal of the rule is achieved.
This functionality revealed an architectural bug which needs to be solved in the future. The documentation stated that rules do not need to be thread-safe and stateless as a new instance of the Rule is used for each file that is processed. However, when running "KtLint.format" the Rule instance is used twice to process a single file. This leaks the state of the child classes of Rule as well as from the Rule class itself. State leakage problems in the child classes are resolved in different rules by manually clearing the state. The state of the Rule class itself is now reset using an additional lifecycle hooks which can not be overridden by child classes.