RuleSetProviders now provide a list of RuleProviders so that the KtLint rule engine can request a new instance of a Rule when needed. KtLint 0.47, supports the old RuleSetProvider and the new RuleSetProviderV2.
Add metadata to RuleSetProvider. This information will be made available via the CLI but can also be used by API integrators. Objective of this metadata is to inform the user about the maintainer of the rule set.

Use property mapper in the disabledRulesProperty to get rid of unwanted spaces around (or inside) the comma separated list as the identifier are not allowed to contain spaces.