Rework rule traversal and fully support non-stateless rules (#1550)
KtLint documentation of prior versions already claimed that rules do not need to be stateless. But according to comments in some rules (for example NoUnusedImportsRule and Identation) some cleanup of state was needed anyway.

When "KtLint.format" autocorrects an error, the Rule instance is reused to process the altered AST for lint violations which have not been fixed. This leaks the state of the Rule instance.

Rules have to be provided using a RuleProvider. Custom RuleSetProviders (CLI) have to provide a set of RuleProviders by implementing RuleSetProviderV2. API consumers have to provide a set of RuleProviders instead of a list of RuleSets. The RuleProvider allows the KtLint rule engine to create a new instance of the Rule whenever needed. Using a RuleRunner as wrapper around a rule, the KtLint rule engine now can track whether the Rule instance has been used for traversal of the AST. If so, the Rule instance is potentially filled with state.

Rules can now stop the traversal of the AST whenever the goal of the rule is achieved.

The RuleSetProvider is now able to store "About" information . This information will be made available via the CLI but can also be used by API integrators. Objective of this metadata is to inform the user about the maintainer of the rule set.

Use property mapper in the disabledRulesProperty to get rid of unwanted spaces around (or inside) the comma separated list as the identifier are not allowed to contain spaces.